spec: expand upon when table processing stops
Previously the spec stated that table processing stops when a matching
flow table entry contains no Go-To table instruction. Additional cases
after the packet has been processed by the last table in the pipeline,
or when no flows match and the switch is not configured to foward to the
next table.