spec: clarify behavior on connection interruption
If a switch loses the connection to the controller it can not send
messages to the controller. Therefore it must either enter fail secure
or fail standalone mode immediately. It may optionally transition from
secure to standalone after a pre-determined period of time or after a
number of failed attempts to contact a controller.