spec: rename set-output-port back to output
The renaming to set-output-port did not work, so let's revert it
First, it define behaviour based on the action name, implicitely, and
when explaining it to people it did not work. The behaviour will need
to be explicitely specified (in the next commits).
Second, set-output-port was misleading when applied to action lists,
as it can't describe the desirted bahaviour.
Third, keeping the same name as in 1.0 helps backward compatibility.